Lipa na M-Pesa is a mobile payment service offered by Safaricom
Lipa na M-Pesa is a mobile payment service offered by Safaricom, a mobile network operator in Kenya. The service allows customers to pay for goods and services by transferring money from their M-Pesa account to a merchant's M-Pesa account.

A Lipa na M-Pesa plugin is a software tool that allows merchants to integrate the Lipa na M-Pesa service into their website or mobile app. This plugin typically provides a secure and seamless way for customers to make payments using Lipa na M-Pesa, without having to leave the merchant's website or app.

To use the plugin, the merchant typically needs to create an account with Lipa na M-Pesa and obtain an API key or other credentials that allow the plugin to connect to the Lipa na M-Pesa service. The plugin then provides a payment form or other user interface that allows customers to enter their M-Pesa details and complete the transaction.